Move FakeRESTClient to client/fake.go
